work through the structs for work and task. Added graphql calls from task and person to work. Changed effort to i32 to keep things simple.

+++ migrations/2021-09-11-192223_add_base_models/up.sql
@@ -279,7 +279,7 @@ CREATE TABLE IF NOT EXISTS tasks (
     target_completion_date TIMESTAMP NOT NULL,
 
     task_status task_status NOT NULL DEFAULT 'planning',
-    effort FLOAT NOT NULL,
+    effort INT NOT NULL,
 
     completed_date TIMESTAMP,
     created_at TIMESTAMP NOT NULL DEFAULT NOW(),
